requirements.txt
anyio==3.7.0
colorama==0.4.6
fastapi==0.98.0
ffmpeg-python==0.2.0
future==0.18.3
idna==3.4
keyboard==0.13.5
PyAudio==0.2.13
pydantic==1.10.9
six==1.16.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.3
whisper==1.1.10
